3948 add transformAggsToAttributeListParam, and support graphql
